requirements.txt
fastapi
uvicorn
joblib
numpy
scikit-learn


requests
pytest
httpx
boto3

python-dotenv
